splash screen for checking authentication
