+++ lib/main.dart
@@ -16,7 +16,8 @@ class MyApp extends StatelessWidget {
         primaryColor: kPrimaryColor,
         scaffoldBackgroundColor: Colors.white,
       ),
-      home: WelcomeScreen(), //TODO: Change based on logged in status
+      // home: WelcomeScreen(), //TODO: Change based on logged in status
+      home: SplashScreen(),
     );
   }
 }
